fix(security): satisfy phpmd/CI for #315
Split ApiRateLimitGraphQlResolverTest (31 methods, exceeded phpmd
TooManyMethods max of 25) into a mutation-behavior class and an
ApiRateLimitGraphQlResolverEdgeCasesTest sibling, hoisting the shared
createGraphQlRequest helper and the GraphQL endpoint constant into
RateLimitClientTestCase. All 30 tests preserved.
