+++ tests/Unit/Shared/Application/Resolver/RateLimit/ApiRateLimitGraphQlResolverTest.php
@@ -8,12 +8,9 @@ use App\Shared\Application\Resolver\RateLimit\ApiRateLimitGraphQlResolver;
 use App\User\Domain\Entity\PendingTwoFactor;
 use App\User\Domain\Repository\PendingTwoFactorRepositoryInterface;
 use DateTimeImmutable;
-use Symfony\Component\HttpFoundation\Request;
 
 final class ApiRateLimitGraphQlResolverTest extends RateLimitClientTestCase
 {
-    private const ENDPOINT = '/api/graphql';
-
     public function testSignInMutationProducesIpAndEmailLimiters(): void
     {
         $email = $this->faker->email();
@@ -136,55 +133,24 @@ final class ApiRateLimitGraphQlResolverTest extends RateLimitClientTestCase
         self::assertNotContains('signout_all', $names);
     }
 
-    public function testNonGraphQlPathReturnsNoTargets(): void
+    public function testSignOutAllMutationProducesUserLimiterForAuthenticatedRequest(): void
     {
-        $request = Request::create('/api/users', 'POST');
-
-        self::assertSame([], $this->createGraphQlResolver()->resolve($request));
-    }
-
-    public function testGraphQlGetRequestReturnsNoTargets(): void
-    {
-        $request = Request::create(self::ENDPOINT, 'GET');
-
-        self::assertSame([], $this->createGraphQlResolver()->resolve($request));
-    }
-
-    public function testNonSensitiveMutationReturnsNoTargets(): void
-    {
-        $request = $this->createGraphQlRequest('mutation { createUser(input: {}) { id } }');
-
-        self::assertSame([], $this->createGraphQlResolver()->resolve($request));
-    }
-
-    public function testInvalidJsonBodyReturnsNoTargets(): void
-    {
-        $request = Request::create(
-            self::ENDPOINT,
-            'POST',
-            [],
-            [],
-            [],
-            ['CONTENT_TYPE' => 'application/json'],
-            'not-json'
+        $userId = $this->faker->uuid();
+        $jwtConverter = $this->createMock(
+            \App\Shared\Application\Converter\JwtTokenConverterInterface::class
         );
+        $jwtConverter->method('decode')->willReturn($this->buildValidPayload(['sub' => $userId]));
 
-        self::assertSame([], $this->createGraphQlResolver()->resolve($request));
-    }
+        $request = $this->createGraphQlRequest('mutation { signOutAll(input: {}) { id } }');
+        $request->headers->set('Authorization', 'Bearer ' . $this->faker->sha256());
 
-    public function testMissingQueryFieldReturnsNoTargets(): void
-    {
-        $request = Request::create(
-            self::ENDPOINT,
-            'POST',
-            [],
-            [],
-            [],
-            ['CONTENT_TYPE' => 'application/json'],
-            json_encode(['variables' => ['input' => []]], JSON_THROW_ON_ERROR)
+        $resolver = $this->createGraphQlResolver(
+            $this->createClientIdentityResolver($jwtConverter)
         );
+        $byName = array_column($resolver->resolve($request), 'key', 'name');
 
-        self::assertSame([], $this->createGraphQlResolver()->resolve($request));
+        self::assertSame('user:' . $userId, $byName['signout_all']);
+        self::assertArrayNotHasKey('signout', $byName);
     }
 
     public function testMultipleSensitiveMutationsAreAllThrottled(): void
@@ -197,82 +163,6 @@ final class ApiRateLimitGraphQlResolverTest extends RateLimitClientTestCase
 
         self::assertContains('refresh_token', $names);
         self::assertContains('password_reset_confirm', $names);
-    }
-
-    public function testScalarJsonBodyReturnsNoTargets(): void
-    {
-        $request = Request::create(
-            self::ENDPOINT,
-            'POST',
-            [],
-            [],
-            [],
-            ['CONTENT_TYPE' => 'application/json'],
-            '"signIn"'
-        );
-
-        self::assertSame([], $this->createGraphQlResolver()->resolve($request));
-    }
-
-    public function testNullJsonBodyReturnsNoTargets(): void
-    {
-        $request = Request::create(
-            self::ENDPOINT,
-            'POST',
-            [],
-            [],
-            [],
-            ['CONTENT_TYPE' => 'application/json'],
-            'null'
-        );
-
-        self::assertSame([], $this->createGraphQlResolver()->resolve($request));
-    }
-
-    public function testNonArrayVariablesAreIgnored(): void
-    {
-        $request = Request::create(
-            self::ENDPOINT,
-            'POST',
-            [],
-            [],
-            [],
-            ['CONTENT_TYPE' => 'application/json'],
-            json_encode(
-                ['query' => 'mutation { signIn(input: {}) { id } }', 'variables' => 'not-an-array'],
-                JSON_THROW_ON_ERROR
-            )
-        );
-
-        $byName = array_column($this->createGraphQlResolver()->resolve($request), 'key', 'name');
-
-        self::assertSame(['signin_ip'], array_keys($byName));
-    }
-
-    public function testEmailFromVariablesIsLowercasedAndTrimmed(): void
-    {
-        $request = $this->createGraphQlRequest(
-            'mutation($input: signInInput!) { signIn(input: $input) { user { id } } }',
-            ['input' => ['email' => "  USER@Example.COM\t"]],
-            '198.51.100.6'
-        );
-
-        $byName = array_column($this->createGraphQlResolver()->resolve($request), 'key', 'name');
-
-        self::assertSame('email:user@example.com', $byName['signin_email']);
-    }
-
-    public function testTopLevelVariablesAreUsedWhenInputKeyMissing(): void
-    {
-        $request = $this->createGraphQlRequest(
-            'mutation($email: String!) { signIn(email: $email) { user { id } } }',
-            ['email' => 'top.level@example.com'],
-            '198.51.100.7'
-        );
-
-        $byName = array_column($this->createGraphQlResolver()->resolve($request), 'key', 'name');
-
-        self::assertSame('email:top.level@example.com', $byName['signin_email']);
     }
 
     public function testDuplicateMutationOccurrencesProduceSingleTargetSet(): void
@@ -357,20 +247,6 @@ final class ApiRateLimitGraphQlResolverTest extends RateLimitClientTestCase
         self::assertArrayNotHasKey('twofa_verification_user', $byName);
     }
 
-    public function testEmptyEmailStringIsTreatedAsAbsent(): void
-    {
-        $request = $this->createGraphQlRequest(
-            'mutation($input: signInInput!) { signIn(input: $input) { user { id } } }',
-            ['input' => ['email' => '']],
-            '198.51.100.44'
-        );
-
-        $byName = array_column($this->createGraphQlResolver()->resolve($request), 'key', 'name');
-
-        self::assertArrayHasKey('signin_ip', $byName);
-        self::assertArrayNotHasKey('signin_email', $byName);
-    }
-
     public function testCompleteTwoFactorWithoutSessionDoesNotQueryRepository(): void
     {
         $repository = $this->createMock(PendingTwoFactorRepositoryInterface::class);
@@ -391,123 +267,5 @@ final class ApiRateLimitGraphQlResolverTest extends RateLimitClientTestCase
 
         self::assertArrayHasKey('twofa_verification_ip', $byName);
         self::assertArrayNotHasKey('twofa_verification_user', $byName);
-    }
-
-    public function testEmailIsResolvedRegardlessOfItsPositionInInput(): void
-    {
-        $request = $this->createGraphQlRequest(
-            'mutation($input: signInInput!) { signIn(input: $input) { user { id } } }',
-            ['input' => ['password' => 'secret', 'email' => 'positioned@example.com']],
-            '198.51.100.47'
-        );
-
-        $byName = array_column($this->createGraphQlResolver()->resolve($request), 'key', 'name');
-
-        self::assertSame('email:positioned@example.com', $byName['signin_email']);
-    }
-
-    public function testNonArrayInputKeyIsIgnored(): void
-    {
-        $request = Request::create(
-            self::ENDPOINT,
-            'POST',
-            [],
-            [],
-            [],
-            ['REMOTE_ADDR' => '198.51.100.46', 'CONTENT_TYPE' => 'application/json'],
-            json_encode(
-                [
-                    'query' => 'mutation { signIn(input: {}) { id } }',
-                    'variables' => ['input' => 'not-an-array'],
-                ],
-                JSON_THROW_ON_ERROR
-            )
-        );
-
-        $byName = array_column($this->createGraphQlResolver()->resolve($request), 'key', 'name');
-
-        self::assertSame(['signin_ip'], array_keys($byName));
-    }
-
-    public function testNonStringQueryFieldReturnsNoTargets(): void
-    {
-        $request = Request::create(
-            self::ENDPOINT,
-            'POST',
-            [],
-            [],
-            [],
-            ['CONTENT_TYPE' => 'application/json'],
-            json_encode(['query' => 12345], JSON_THROW_ON_ERROR)
-        );
-
-        self::assertSame([], $this->createGraphQlResolver()->resolve($request));
-    }
-
-    public function testNestedInputObjectIsDecodedAssociatively(): void
-    {
-        $request = Request::create(
-            self::ENDPOINT,
-            'POST',
-            [],
-            [],
-            [],
-            ['CONTENT_TYPE' => 'application/json'],
-            json_encode(
-                [
-                    'query' => 'mutation { signIn(input: {}) { id } }',
-                    'variables' => ['input' => ['email' => 'nested@example.com']],
-                ],
-                JSON_THROW_ON_ERROR
-            )
-        );
-
-        $byName = array_column($this->createGraphQlResolver()->resolve($request), 'key', 'name');
-
-        self::assertSame('email:nested@example.com', $byName['signin_email']);
-    }
-
-    public function testSignOutAllMutationProducesUserLimiterForAuthenticatedRequest(): void
-    {
-        $userId = $this->faker->uuid();
-        $jwtConverter = $this->createMock(
-            \App\Shared\Application\Converter\JwtTokenConverterInterface::class
-        );
-        $jwtConverter->method('decode')->willReturn($this->buildValidPayload(['sub' => $userId]));
-
-        $request = $this->createGraphQlRequest('mutation { signOutAll(input: {}) { id } }');
-        $request->headers->set('Authorization', 'Bearer ' . $this->faker->sha256());
-
-        $resolver = $this->createGraphQlResolver(
-            $this->createClientIdentityResolver($jwtConverter)
-        );
-        $byName = array_column($resolver->resolve($request), 'key', 'name');
-
-        self::assertSame('user:' . $userId, $byName['signout_all']);
-        self::assertArrayNotHasKey('signout', $byName);
-    }
-
-    /**
-     * @param array<string, array<string, scalar|null>|scalar|null> $variables
-     */
-    private function createGraphQlRequest(
-        string $query,
-        array $variables = [],
-        string $clientIp = '203.0.113.7'
-    ): Request {
-        $body = ['query' => $query];
-        if ($variables !== []) {
-            $body['variables'] = $variables;
-        }
-
-        return Request::create(
-            self::ENDPOINT,
-            'POST',
-            [],
-            [],
-            [],
-            ['REMOTE_ADDR' => $clientIp, 'CONTENT_TYPE' => 'application/json'],
-            json_encode($body, JSON_THROW_ON_ERROR)
-        );
     }
 }
